+++ src/index.ts
@@ -1,14 +1,17 @@
 import { events, Product, Response, Stock, Order, OrderItem, OrderPart, User } from '@/types'
-import { init, updateToken, updateInstanceUrl } from '@/api'
+import { init, resetConfig, updateToken, updateInstanceUrl } from '@/api'
 import { isError } from '@/util'
-import { fetchProducts, fetchProductsStock, getOrderDetails, getProfile, updateOrderStatus } from '@/modules'
+import { fetchProducts, fetchProductsGroupedBy, fetchProductsGroupedByParent, fetchProductsStock, getOrderDetails, getProfile, updateOrderStatus } from '@/modules'
 
 export {
   getOrderDetails,
   updateOrderStatus,
   fetchProducts,
+  fetchProductsGroupedBy,
+  fetchProductsGroupedByParent,
   init,
   isError,
+  resetConfig,
   updateToken,
   updateInstanceUrl,
   fetchProductsStock,