+++ src/modules/index.ts
@@ -1,10 +1,12 @@
 import { getOrderDetails, updateOrderStatus } from '@/modules/order'
-import { fetchProducts } from '@/modules/product'
+import { fetchProducts, fetchProductsGroupedBy, fetchProductsGroupedByParent } from '@/modules/product'
 import { getProfile } from '@/modules/user'
 import { fetchProductsStock } from '@/modules/stock'
 
 export {
   fetchProducts,
+  fetchProductsGroupedBy,
+  fetchProductsGroupedByParent,
   fetchProductsStock,
   getOrderDetails,
   getProfile,